made forge version also work
