+++ common/src/main/kotlin/org/valkyrienskies/clockwork/content/curiosities/tools/gravitron/tool/GrabTool.kt
@@ -18,7 +18,7 @@ import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.Gravitro
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronGrabPacket
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronItem.Companion.GravitronState
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronItem.Companion.getState
-import org.valkyrienskies.clockwork.util.ClockworkUtils.readVec3
+import org.valkyrienskies.clockwork.util.ClockworkUtil.readVec3
 import org.valkyrienskies.core.api.ships.LoadedServerShip
 import org.valkyrienskies.core.api.ships.ServerShip
 import org.valkyrienskies.mod.common.getShipManagingPos
