+++ common/src/main/kotlin/org/valkyrienskies/clockwork/platform/SharedValues.kt
@@ -5,6 +5,7 @@ import com.tterrag.registrate.util.entry.BlockEntityEntry
 import com.tterrag.registrate.util.entry.EntityEntry
 import dev.architectury.injectables.annotations.ExpectPlatform
 import net.minecraft.world.item.BlockItem
+import org.valkyrienskies.clockwork.content.contraptions.smart_propeller.SuperContraptionEntity
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronHandler
 import org.valkyrienskies.clockwork.content.curiosities.tools.wanderwand.WanderWandClusterRenderer
 import org.valkyrienskies.clockwork.content.kinetics.sequenced_seat.SequencedSeatEntity
@@ -35,6 +36,13 @@ object SharedValues {
     @get:ExpectPlatform
     @JvmStatic
     val sequencedSeat: EntityEntry<SequencedSeatEntity>
+        get() {
+            throw AssertionError()
+        }
+
+    @get:ExpectPlatform
+    @JvmStatic
+    val superContraption: EntityEntry<SuperContraptionEntity>
         get() {
             throw AssertionError()
         }
