+++ fabric/src/main/java/org/valkyrienskies/clockwork/platform/fabric/SharedValuesImpl.java
@@ -6,6 +6,7 @@ import com.tterrag.registrate.util.entry.EntityEntry;
 import net.fabricmc.fabric.api.client.rendering.v1.BuiltinItemRendererRegistry;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
+import org.valkyrienskies.clockwork.content.contraptions.smart_propeller.SuperContraptionEntity;
 import org.valkyrienskies.clockwork.content.curiosities.tools.wanderwand.WanderWandClusterRenderer;
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronHandler;
 import org.valkyrienskies.clockwork.content.kinetics.sequenced_seat.SequencedSeatEntity;
@@ -34,6 +35,10 @@ public class SharedValuesImpl {
 
     public static EntityEntry<SequencedSeatEntity> getSequencedSeat() {
         return (EntityEntry) FabricClockworkEntities.SEQUENCED_SEAT;
+    }
+
+    public static EntityEntry<SuperContraptionEntity> getSuperContraption() {
+        return (EntityEntry) FabricClockworkEntities.SUPER_CONTRAPTION_ENTITY;
     }
 
     public static BlockEntityEntry<BallastBlockEntity> getBallast() {
