+++ forge/src/main/java/org/valkyrienskies/clockwork/forge/ForgeClockworkEntities.java
@@ -2,6 +2,7 @@ package org.valkyrienskies.clockwork.forge;
 
 import com.simibubi.create.content.contraptions.AbstractContraptionEntity;
 import com.simibubi.create.content.contraptions.actors.seat.SeatEntity;
+import com.simibubi.create.content.contraptions.render.ContraptionEntityRenderer;
 import com.simibubi.create.foundation.data.CreateEntityBuilder;
 import com.tterrag.registrate.util.entry.EntityEntry;
 import com.tterrag.registrate.util.nullness.NonNullConsumer;
@@ -14,6 +15,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
 import org.valkyrienskies.clockwork.ClockworkLang;
 import org.valkyrienskies.clockwork.ClockworkMod;
+import org.valkyrienskies.clockwork.content.contraptions.smart_propeller.SuperContraptionEntity;
 import org.valkyrienskies.clockwork.platform.entity.ForgeSequencedSeatEntity;
 
 public class ForgeClockworkEntities {
@@ -29,6 +31,12 @@ public class ForgeClockworkEntities {
                     false,
                     true,
                     ForgeSequencedSeatEntity::build
+            ).register();
+
+    public static final EntityEntry<SuperContraptionEntity> SUPER_CONTRAPTION_ENTITY = contraption("super_contraption",
+            SuperContraptionEntity::new,
+            () -> ContraptionEntityRenderer::new,
+            20, 40, false
             ).register();
 
     private static <T extends Entity> CreateEntityBuilder<T, ?> contraption(String name, EntityType.EntityFactory<T> factory,
